requirements.txt
rich>=13.0.0
prompt-toolkit>=3.0.0
httpx>=0.25.0
openai>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0
numpy>=1.24.0
chromadb>=0.4.0
pyyaml>=6.0.0
jinja2>=3.1.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0
cryptography>=41.0.0
pydantic>=2.0.0
click>=8.1.0
watchdog>=3.0.0
docker>=7.0.0
gitpython>=3.1.0
python-dotenv>=1.0.0
cachetools>=5.3.0
tenacity>=8.2.0
pygments>=2.17.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
